fix lại giao diện và trang quản lí thêm sản phẩm

+++ database/migrations/2026_09_10_203617_add_ward_to_orders_table.php
@@ -9,14 +9,18 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->string('ward')->nullable()->after('district');
+            if (!Schema::hasColumn('orders', 'ward')) {
+                $table->string('ward')->nullable()->after('district');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->dropColumn('ward');
+            if (Schema::hasColumn('orders', 'ward')) {
+                $table->dropColumn('ward');
+            }
         });
     }
 };
